telegram-bot/requirements.txt
python-telegram-bot==20.7
requests==2.31.0
python-dotenv==1.0.0
aiohttp==3.9.1
# Добавленные зависимости для улучшенной архитектуры
pydantic==2.5.0  # Для валидации данных
structlog==23.2.0  # Для структурированного логирования
tenacity==8.2.3  # Для retry логики
prometheus-client==0.19.0  # Для метрик (опционально)
# Зависимости для тестирования
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
pytest-mock==3.12.0 